requirements.txt
azure-ai-textanalytics==5.3.0
azure-core==1.29.4
python-dotenv==1.0.1
